modification added for vtk legacy
